+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Version 1.0.0-alpha3
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Refactoring store view code handling
+* VariantSubject now extends AbstractProductSubject
+* Add Robo.li composer dependeny + task configuration
+* ProductVariantProcessorInterface now extends ProductProcessorInterface
+
 # Version 1.0.0-alpha2
 
 ## Bugfixes